requirements.txt
dataclasses
datasets
transformers
trl
accelerate
awscli
peft
scipy
bitsandbytes
deepspeed

rich
asgi-correlation-id
python-dotenv